refactor: update

+++ src/actions/map.ts
@@ -1,20 +1,24 @@
 'use server'
 
 import { calculateMapBounds, calculateScoreFactor } from '../utils/game.js'
+import { getCountryFromCoordinates } from '../utils/map.js'
 
 import { createClient } from '../utils/supabase/server.js'
 
-import type { Location, Map, Profile } from '../types/index.js'
 import { createMapValidation } from '../utils/validations/map.js'
 
-export const getCommunityMaps = async () => {
+import type { Coords, Location, Map, Profile } from '../types/index.js'
+
+export const getRandomCommunityMaps = async (limit: number) => {
   'use server'
   const supabase = createClient()
 
   const { data, error } = await supabase
-    .from('maps')
+    .rpc('get_random_maps', {
+      p_type: 'community',
+      p_count: limit,
+    })
     .select('*')
-    .eq('type', 'community')
     .returns<Map[]>()
 
   return {
@@ -23,7 +27,45 @@ export const getCommunityMaps = async () => {
   }
 }
 
-export const getOfficialMaps = async () => {
+export const getRandomOfficialMaps = async (limit: number) => {
+  'use server'
+  const supabase = createClient()
+
+  const { data, error } = await supabase
+    .rpc('get_random_maps', {
+      p_type: 'official',
+      p_count: limit,
+    })
+    .select('*')
+    .returns<Map[]>()
+
+  return {
+    data,
+    error: error?.message ?? null,
+  }
+}
+
+const PAGE_PER_MAPS = 20
+
+export const getCommunityMaps = async (page: number) => {
+  'use server'
+  const supabase = createClient()
+
+  const { data, error } = await supabase
+    .from('maps')
+    .select('*')
+    .eq('type', 'community')
+    .eq('is_public', true)
+    .range(page * PAGE_PER_MAPS, (page + 1) * PAGE_PER_MAPS - 1)
+    .returns<Map[]>()
+
+  return {
+    data,
+    error: error?.message ?? null,
+  }
+}
+
+export const getOfficialMaps = async (page: number) => {
   'use server'
   const supabase = createClient()
 
@@ -31,6 +73,8 @@ export const getOfficialMaps = async () => {
     .from('maps')
     .select('*')
     .eq('type', 'official')
+    .eq('is_public', true)
+    .range(page * PAGE_PER_MAPS, (page + 1) * PAGE_PER_MAPS - 1)
     .returns<Map[]>()
 
   return {
@@ -55,20 +99,11 @@ export const getMap = async (id: string) => {
   }
 }
 
-export const createOfficialMap = async () => {
-  'use server'
-  const supabase = createClient()
-
-  await supabase.from('maps').insert({
-    type: 'official',
-  })
-}
-
 export const createCommunityMap = async (data: {
   name: string
   description: string | null
   creator: string
-  public: boolean
+  isPublic: boolean
 }) => {
   'use server'
   const supabase = createClient()
@@ -88,12 +123,12 @@ export const createCommunityMap = async (data: {
       name: validated.data.name,
       description: validated.data.description,
       creator: validated.data.creator,
-      public: validated.data.public,
+      is_public: validated.data.isPublic,
 
       //
       locations_count: 0,
       score_factor: 0,
-      bounds: { min: { lat: 0, lng: 0 }, max: { lat: 0, lng: 0 } },
+      bounds: null,
     })
     .select()
     .single<Map>()
@@ -110,7 +145,7 @@ export const updateMap = async (
   id: string,
   data: {
     name?: string
-    locations?: Omit<Location, 'id' | 'map_id' | 'user_id'>[]
+    locations?: Coords[]
   },
 ) => {
   'use server'
@@ -182,37 +217,39 @@ export const updateMap = async (
       newLocs.every((n) => n.pano_id !== o.pano_id),
     )
 
-    console.log('oldLocs', oldLocs)
-    console.log('newLocs', newLocs)
+    if (removing.length > 0) {
+      for (const loc of removing) {
+        const { error: err } = await supabase
+          .from('locations')
+          .delete()
+          .eq('map_id', mapData.id)
+          .eq('pano_id', loc.pano_id)
 
-    console.log('adding', adding)
-    console.log('removing', removing)
-
-    for (const loc of removing) {
-      const { error: err } = await supabase
-        .from('locations')
-        .delete()
-        .eq('map_id', mapData.id)
-        .eq('pano_id', loc.pano_id)
-
-      if (err)
-        return {
-          data: null,
-          error: err?.message ?? null,
-        }
+        if (err)
+          return {
+            data: null,
+            error: err?.message ?? null,
+          }
+      }
     }
 
-    for (const loc of adding) {
-      const { error: err } = await supabase.from('locations').insert({
-        map_id: mapData.id,
-        user_id: uData.user.id,
-        ...loc,
-      })
+    if (adding.length > 0) {
+      const { error: insertedErr } = await supabase.from('locations').insert(
+        adding.map((loc) => ({
+          map_id: mapData.id,
+          user_id: uData.user.id,
+          streak_location_code: getCountryFromCoordinates({
+            lat: loc.lat,
+            lng: loc.lng,
+          }),
+          ...loc,
+        })),
+      )
 
-      if (err)
+      if (insertedErr)
         return {
           data: null,
-          error: err?.message ?? null,
+          error: insertedErr?.message ?? null,
         }
     }
   }
@@ -253,11 +290,74 @@ export const updateMap = async (
     .select()
     .single<Map>()
 
-  console.log('map has updated', updatedData, updatedErr)
-
   return {
     data: updatedData,
     error: updatedErr?.message ?? null,
+  }
+}
+
+export const deleteMap = async (id: string) => {
+  'use server'
+  const supabase = createClient()
+
+  const { data: uData, error: uErr } = await supabase.auth.getUser()
+
+  if (!uData.user || uErr)
+    return {
+      data: null,
+      error: uErr?.message ?? null,
+    }
+
+  const { data: pData, error: pErr } = await supabase
+    .from('profiles')
+    .select('*')
+    .eq('id', uData.user.id)
+    .single<Profile>()
+
+  if (!pData || pErr)
+    return {
+      data: null,
+      error: pErr?.message ?? null,
+    }
+
+  const { data: mapData, error: mErr } = await supabase
+    .from('maps')
+    .select('*')
+    .eq('id', id)
+    .single<Map>()
+
+  if (!mapData || mErr)
+    return {
+      data: null,
+      error: mErr?.message ?? null,
+    }
+
+  if (mapData.creator !== uData.user.id)
+    return {
+      data: null,
+      error: mErr,
+    }
+
+  if (mapData.type === 'official' && !pData.is_admin)
+    return {
+      data: null,
+      error: null,
+    }
+
+  const { error: deletedErr } = await supabase
+    .from('maps')
+    .delete()
+    .eq('id', mapData.id)
+
+  if (deletedErr)
+    return {
+      data: null,
+      error: deletedErr?.message ?? null,
+    }
+
+  return {
+    data: true,
+    error: null,
   }
 }
